Adding options for viewing the data in the form of a table
