+++ templates/admin4.html
@@ -28,6 +28,7 @@
     <button onclick="showForm('createForm')">Create New User</button>
     <button onclick="showForm('updateForm')">Update User</button>
     <button onclick="showForm('deleteForm')">Delete User</button>
+    <button onclick="location.href='/view_users'">View Users</button>
     
     <form id="createForm" style="display:none" method="post">
         <input type="hidden" name="action" value="create">
