+++ templates/assign_role.html
@@ -23,11 +23,12 @@
     </script>
 </head>
 <body>
-    <h1>Assign Roles to Users</h1>
+    <h1 class="page-title">User Management</h1>
     
     <button onclick="showForm('createForm')">Create New User</button>
     <button onclick="showForm('updateForm')">Update User</button>
     <button onclick="showForm('deleteForm')">Delete User</button>
+    <button onclick="location.href='/view_users'">View Users</button>
     
     <form id="createForm" style="display:none" method="post">
         <input type="hidden" name="action" value="create">
@@ -101,8 +102,6 @@
         <button type="button" onclick="location.href='/assign_role'">Cancel</button>
     </form>
     {% endif %}
-
-    
 
 </body>
 </html>
